Take amp-list routes out of app.js into its own routing file (#21214)
* Break amp-list routes out into own routing file

* Refactor amp-list manual tests and routes out to their own file and folder

* Update other usages of enableCors
