+++ build-system/amp-cors.js
@@ -14,6 +14,69 @@
  * limitations under the License.
  */
 
+/**
+ * In practice this would be *.ampproject.org and the publishers
+ * origin. Please see AMP CORS docs for more details:
+ *    https://goo.gl/F6uCAY
+ * @type {RegExp}
+ */
+const ORIGIN_REGEX = new RegExp('^http://localhost:8000|' +
+'^https?://.+\.herokuapp\.com');
+
+/**
+* In practice this would be the publishers origin.
+* Please see AMP CORS docs for more details:
+*    https://goo.gl/F6uCAY
+* @type {RegExp}
+*/
+const SOURCE_ORIGIN_REGEX = new RegExp('^http://localhost:8000|' +
+'^https?://.+\.herokuapp\.com');
+
+function assertCors(req, res, opt_validMethods, opt_exposeHeaders,
+  opt_ignoreMissingSourceOrigin) {
+  // Allow disable CORS check (iframe fixtures have origin 'about:srcdoc').
+  if (req.query.cors == '0') {
+    return;
+  }
+
+  const validMethods = opt_validMethods || ['GET', 'POST', 'OPTIONS'];
+  const invalidMethod = req.method + ' method is not allowed. Use POST.';
+  const invalidOrigin = 'Origin header is invalid.';
+  const invalidSourceOrigin = '__amp_source_origin parameter is invalid.';
+  const unauthorized = 'Unauthorized Request';
+  let origin;
+
+  if (validMethods.indexOf(req.method) == -1) {
+    res.statusCode = 405;
+    res.end(JSON.stringify({message: invalidMethod}));
+    throw invalidMethod;
+  }
+
+  if (req.headers.origin) {
+    origin = req.headers.origin;
+    if (!ORIGIN_REGEX.test(req.headers.origin)) {
+      res.statusCode = 500;
+      res.end(JSON.stringify({message: invalidOrigin}));
+      throw invalidOrigin;
+    }
+
+    if (!opt_ignoreMissingSourceOrigin &&
+        !SOURCE_ORIGIN_REGEX.test(req.query.__amp_source_origin)) {
+      res.statusCode = 500;
+      res.end(JSON.stringify({message: invalidSourceOrigin}));
+      throw invalidSourceOrigin;
+    }
+  } else if (req.headers['amp-same-origin'] == 'true') {
+    origin = getUrlPrefix(req);
+  } else {
+    res.statusCode = 401;
+    res.end(JSON.stringify({message: unauthorized}));
+    throw unauthorized;
+  }
+
+  enableCors(req, res, origin, opt_exposeHeaders);
+}
+
 function enableCors(req, res, origin, opt_exposeHeaders) {
   res.setHeader('Access-Control-Allow-Credentials', 'true');
 
@@ -33,4 +96,10 @@ function enableCors(req, res, origin, opt_exposeHeaders) {
   }
 }
 
-module.exports = enableCors;
+function getUrlPrefix(req) {
+  return req.protocol + '://' + req.headers.host;
+}
+
+module.exports = {
+  enableCors, assertCors,
+};
