+++ build-system/app.js
@@ -23,8 +23,8 @@ const app = require('express')();
 const bacon = require('baconipsum');
 const BBPromise = require('bluebird');
 const bodyParser = require('body-parser');
+const cors = require('./amp-cors');
 const devDashboard = require('./app-index/index');
-const enableCors = require('./amp-cors');
 const formidable = require('formidable');
 const fs = BBPromise.promisifyAll(require('fs'));
 const jsdom = require('jsdom');
@@ -32,7 +32,6 @@ const multer = require('multer');
 const path = require('path');
 const request = require('request');
 const pc = process;
-const countries = require('../examples/countries.json');
 const runVideoTestBench = require('./app-video-testbench');
 const {
   recaptchaFrameRequestHandler,
@@ -46,6 +45,7 @@ const upload = multer();
 app.use(bodyParser.text());
 app.use('/amp4test', require('./amp4test').app);
 app.use('/analytics', require('./routes/analytics'));
+app.use('/list/', require('./routes/list'));
 
 // Append ?csp=1 to the URL to turn on the CSP header.
 // TODO: shall we turn on CSP all the time?
@@ -247,26 +247,8 @@ app.use('/api/echo/post', (req, res) => {
   res.end(req.body);
 });
 
-/**
- * In practice this would be *.ampproject.org and the publishers
- * origin. Please see AMP CORS docs for more details:
- *    https://goo.gl/F6uCAY
- * @type {RegExp}
- */
-const ORIGIN_REGEX = new RegExp('^http://localhost:8000|' +
-    '^https?://.+\.herokuapp\.com');
-
-/**
- * In practice this would be the publishers origin.
- * Please see AMP CORS docs for more details:
- *    https://goo.gl/F6uCAY
- * @type {RegExp}
- */
-const SOURCE_ORIGIN_REGEX = new RegExp('^http://localhost:8000|' +
-    '^https?://.+\.herokuapp\.com');
-
 app.use('/form/html/post', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
 
   const form = new formidable.IncomingForm();
   form.parse(req, (err, fields) => {
@@ -288,14 +270,14 @@ app.use('/form/html/post', (req, res) => {
 
 
 app.use('/form/redirect-to/post', (req, res) => {
-  assertCors(req, res, ['POST'], ['AMP-Redirect-To']);
+  cors.assertCors(req, res, ['POST'], ['AMP-Redirect-To']);
   res.setHeader('AMP-Redirect-To', 'https://google.com');
   res.end('{}');
 });
 
 
 app.use('/form/echo-json/post', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   const form = new formidable.IncomingForm();
   form.parse(req, (err, fields) => {
     res.setHeader('Content-Type', 'application/json; charset=utf-8');
@@ -307,7 +289,7 @@ app.use('/form/echo-json/post', (req, res) => {
 });
 
 app.use('/form/json/poll1', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   const form = new formidable.IncomingForm();
   form.parse(req, () => {
     res.setHeader('Content-Type', 'application/json');
@@ -330,7 +312,7 @@ app.use('/form/json/poll1', (req, res) => {
 });
 
 app.post('/form/json/upload', upload.fields([{name: 'myFile'}]), (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
 
   /** @type {!Array<!File>|undefined} */
   const myFile = req.files['myFile'];
@@ -359,7 +341,7 @@ app.use('/form/search-html/get', (req, res) => {
 
 
 app.use('/form/search-json/get', (req, res) => {
-  assertCors(req, res, ['GET']);
+  cors.assertCors(req, res, ['GET']);
   res.json({
     term: req.query.term,
     additionalFields: req.query.additionalFields,
@@ -373,7 +355,7 @@ const autosuggestLanguages = ['ActionScript', 'AppleScript', 'Asp', 'BASIC',
   'Ruby', 'Scala', 'Scheme'];
 
 app.use('/form/autosuggest/query', (req, res) => {
-  assertCors(req, res, ['GET']);
+  cors.assertCors(req, res, ['GET']);
   const MAX_RESULTS = 4;
   const query = req.query.q;
   if (!query) {
@@ -391,7 +373,7 @@ app.use('/form/autosuggest/query', (req, res) => {
 });
 
 app.use('/form/autosuggest/search', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   const form = new formidable.IncomingForm();
   form.parse(req, function(err, fields) {
     res.json({
@@ -402,7 +384,7 @@ app.use('/form/autosuggest/search', (req, res) => {
 });
 
 app.use('/form/verify-search-json/post', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   const form = new formidable.IncomingForm();
   form.parse(req, (err, fields) => {
     res.setHeader('Content-Type', 'application/json; charset=utf-8');
@@ -470,7 +452,7 @@ function proxyToAmpProxy(req, res, mode) {
     if (inabox) {
       // Allow CORS requests for A4A.
       const origin = req.headers.origin || urlPrefix;
-      enableCors(req, res, origin);
+      cors.enableCors(req, res, origin);
     }
     res.status(response.statusCode).send(body);
   });
@@ -682,7 +664,7 @@ app.use('/examples/bind/live-list.amp.html',
     });
 
 app.use('/impression-proxy/', (req, res) => {
-  assertCors(req, res, ['GET']);
+  cors.assertCors(req, res, ['GET']);
   // Fake response with the following optional fields:
   // location: The Url the that server would have sent redirect to w/o ALP
   // tracking_url: URL that should be requested to track click
@@ -698,7 +680,7 @@ app.use('/impression-proxy/', (req, res) => {
 });
 
 app.post('/get-consent-v1/', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   const body = {
     'promptIfUnknown': true,
     'sharedData': {
@@ -710,7 +692,7 @@ app.post('/get-consent-v1/', (req, res) => {
 });
 
 app.post('/get-consent-no-prompt/', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   const body = {};
   res.json(body);
 });
@@ -739,7 +721,7 @@ app.get('/iframe/*', (req, res) => {
 });
 
 app.get('/a4a_template/*', (req, res) => {
-  assertCors(req, res, ['GET'], undefined, true);
+  cors.assertCors(req, res, ['GET'], undefined, true);
   const match = /^\/a4a_template\/([a-z-]+)\/(\d+)$/.exec(req.path);
   if (!match) {
     res.status(404);
@@ -882,7 +864,7 @@ app.get(['/examples/*.html', '/test/manual/*.html'], (req, res, next) => {
 
     if (inabox && req.headers.origin && req.query.__amp_source_origin) {
       // Allow CORS requests for A4A.
-      enableCors(req, res, req.headers.origin);
+      cors.enableCors(req, res, req.headers.origin);
     } else {
       file = replaceUrls(mode, file, '', inabox);
     }
@@ -940,7 +922,7 @@ function elementExtractor(tagName, type) {
 
 // Data for example: http://localhost:8000/examples/bind/xhr.amp.html
 app.use('/bind/form/get', (req, res) => {
-  assertCors(req, res, ['GET']);
+  cors.assertCors(req, res, ['GET']);
   res.json({
     bindXhrResult: 'I was fetched from the server!',
   });
@@ -948,7 +930,7 @@ app.use('/bind/form/get', (req, res) => {
 
 // Data for example: http://localhost:8000/examples/bind/ecommerce.amp.html
 app.use('/bind/ecommerce/sizes', (req, res) => {
-  assertCors(req, res, ['GET']);
+  cors.assertCors(req, res, ['GET']);
   setTimeout(() => {
     const prices = {
       '0': {
@@ -1011,31 +993,9 @@ app.use('/bind/ecommerce/sizes', (req, res) => {
   }, 1000); // Simulate network delay.
 });
 
-app.use('/list/fruit-data/get', (req, res) => {
-  assertCors(req, res, ['GET']);
-  res.json({
-    items: [
-      {name: 'apple', quantity: 47, unitPrice: '0.33'},
-      {name: 'pear', quantity: 538, unitPrice: '0.54'},
-      {name: 'tomato', quantity: 0, unitPrice: '0.23'},
-    ],
-  });
-});
-
-app.use('/list/vegetable-data/get', (req, res) => {
-  assertCors(req, res, ['GET']);
-  res.json({
-    items: [
-      {name: 'cabbage', quantity: 5, unitPrice: '1.05'},
-      {name: 'carrot', quantity: 10, unitPrice: '0.01'},
-      {name: 'brocoli', quantity: 7, unitPrice: '0.02'},
-    ],
-  });
-});
-
 // Simulated subscription entitlement
 app.use('/subscription/:id/entitlements', (req, res) => {
-  assertCors(req, res, ['GET']);
+  cors.assertCors(req, res, ['GET']);
   res.json({
     source: 'local' + req.params.id,
     granted: true,
@@ -1047,7 +1007,7 @@ app.use('/subscription/:id/entitlements', (req, res) => {
 });
 
 app.use('/subscription/pingback', (req, res) => {
-  assertCors(req, res, ['POST']);
+  cors.assertCors(req, res, ['POST']);
   res.json({
     done: true,
   });
@@ -1055,7 +1015,7 @@ app.use('/subscription/pingback', (req, res) => {
 
 // Simulated adzerk ad server and AMP cache CDN.
 app.get('/adzerk/*', (req, res) => {
-  assertCors(req, res, ['GET'], ['AMP-template-amp-creative']);
+  cors.assertCors(req, res, ['GET'], ['AMP-template-amp-creative']);
   const match = /\/(\d+)/.exec(req.path);
   if (!match || !match[1]) {
     res.status(404);
@@ -1237,109 +1197,6 @@ app.get('/dist/ww(.max)?.js', (req, res) => {
   });
 });
 
-/*
- * Infinite scroll related endpoints.
- */
-const randInt = n => {
-  return Math.floor(Math.random() * Math.floor(n));
-};
-
-const squareImgUrl = width => {
-  return `http://picsum.photos/${width}?${randInt(50)}`;
-};
-
-const generateJson = numberOfItems => {
-  const results = [];
-  for (let i = 0; i < numberOfItems; i++) {
-    const imageUrl = squareImgUrl(200);
-    const r = {
-      'title': 'Item ' + randInt(100),
-      imageUrl,
-      'price': randInt(8) + 0.99,
-    };
-    results.push(r);
-  }
-  return results;
-};
-
-const generateResults = (category, count = 2) => {
-  const r = {};
-  const items = [];
-  for (let i = 0; i < count; i++) {
-    const buster = randInt(10000);
-    const item = {};
-    item.src = `https://placeimg.com/600/400/${category}?${buster}`;
-    items.push(item);
-  }
-
-  r.items = items;
-  r['load-more-src'] =
-      `/infinite-scroll-random/${category}?${randInt(10000)}`;
-
-  return r;
-};
-
-app.get('/infinite-scroll-random/:category', function(request, response) {
-  const {category} = request.params;
-  const result = generateResults(category);
-  response.json(result);
-});
-
-app.get('/infinite-scroll-faulty', function(req, res) {
-  const {query} = req;
-  const code = query['code'];
-  const items = generateJson(12);
-  let next = '/infinite-scroll-error';
-  if (code) {
-    next += '?code=' + code;
-  }
-  res.json({items, next});
-});
-
-app.get('/infinite-scroll-error', function(req, res) {
-  const {query} = req;
-  const code = query['code'] || 404;
-  res.status(code);
-  res.json({'msg': code});
-});
-
-app.get('/infinite-scroll', function(req, res) {
-  const {query} = req;
-  const numberOfItems = query['items'] || 10;
-  const pagesLeft = query['left'] || 1;
-  const latency = query['latency'] || 0;
-
-  const items = generateJson(numberOfItems);
-
-  const nextUrl = '/infinite-scroll?items=' +
-    numberOfItems + '&left=' + (pagesLeft - 1) +
-    '&latency=' + latency;
-
-  const randomFalsy = () => {
-    const rand = Math.floor(Math.random() * Math.floor(3));
-    switch (rand) {
-      case 1: return null;
-      case 2: return undefined;
-      case 3: return '';
-      default: return false;
-    }
-  };
-
-  const next = pagesLeft == 0 ? randomFalsy() : nextUrl;
-  const results = next === false ? {items}
-    : {items, next,
-      'loadMoreButtonText': 'test',
-      'loadMoreEndText': 'end',
-    };
-
-  if (latency) {
-    setTimeout(() => res.json(results), latency);
-  } else {
-    res.json(results);
-  }
-});
-
-
 /**
  * Shadow viewer
  */
@@ -1355,29 +1212,6 @@ app.use('/shadow/', (req, res) => {
     src: req.url.replace(/^\//, ''),
     baseHref,
   }));
-});
-
-
-/**
- * Autosuggest endpoint
- */
-app.get('/search/countries', function(req, res) {
-  let filtered = [];
-  if (req.query.hasOwnProperty('q')) {
-    const query = req.query.q.toLowerCase();
-
-    filtered = countries.items
-        .filter(country => country.name.toLowerCase().startsWith(query));
-  }
-
-  const results = {
-    'items': [
-      {
-        'results': filtered,
-      },
-    ],
-  };
-  res.send(results);
 });
 
 /**
@@ -1426,51 +1260,6 @@ function addQueryParam(url, param, value) {
     url += '&' + paramValue;
   }
   return url;
-}
-
-function assertCors(req, res, opt_validMethods, opt_exposeHeaders,
-  opt_ignoreMissingSourceOrigin) {
-  // Allow disable CORS check (iframe fixtures have origin 'about:srcdoc').
-  if (req.query.cors == '0') {
-    return;
-  }
-
-  const validMethods = opt_validMethods || ['GET', 'POST', 'OPTIONS'];
-  const invalidMethod = req.method + ' method is not allowed. Use POST.';
-  const invalidOrigin = 'Origin header is invalid.';
-  const invalidSourceOrigin = '__amp_source_origin parameter is invalid.';
-  const unauthorized = 'Unauthorized Request';
-  let origin;
-
-  if (validMethods.indexOf(req.method) == -1) {
-    res.statusCode = 405;
-    res.end(JSON.stringify({message: invalidMethod}));
-    throw invalidMethod;
-  }
-
-  if (req.headers.origin) {
-    origin = req.headers.origin;
-    if (!ORIGIN_REGEX.test(req.headers.origin)) {
-      res.statusCode = 500;
-      res.end(JSON.stringify({message: invalidOrigin}));
-      throw invalidOrigin;
-    }
-
-    if (!opt_ignoreMissingSourceOrigin &&
-        !SOURCE_ORIGIN_REGEX.test(req.query.__amp_source_origin)) {
-      res.statusCode = 500;
-      res.end(JSON.stringify({message: invalidSourceOrigin}));
-      throw invalidSourceOrigin;
-    }
-  } else if (req.headers['amp-same-origin'] == 'true') {
-    origin = getUrlPrefix(req);
-  } else {
-    res.statusCode = 401;
-    res.end(JSON.stringify({message: unauthorized}));
-    throw unauthorized;
-  }
-
-  enableCors(req, res, origin, opt_exposeHeaders);
 }
