--- test/manual/amp-list-load-more.amp.html
@@ -1,72 +1,0 @@
-<!doctype html>
-<html ⚡>
-<head>
-  <meta charset="utf-8">
-  <title>AMP #0</title>
-  <link rel="canonical" href="amps.html" >
-  <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
-  <link href='https://fonts.googleapis.com/css?family=Questrial' rel='stylesheet' type='text/css'>
-  <style amp-custom>
-    body {
-      font-family: 'Questrial', Arial;
-    }
-    article {
-      display: block;
-      margin: 8px;
-    }
-    amp-list {
-      border: 1px solid green;
-    }
-    [load-more-button], [load-more-failed] {
-      background: gray;
-      color: #fff;
-      cursor: pointer;
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      right: 0;
-      z-index: 2;
-      padding: 16px;
-    }
-    .story-entry {
-      padding: 40px;
-      border-bottom: 1px solid green;
-    }
-  </style>
-  <style amp-boilerplate>body{-webkit-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-moz-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-ms-animation:-amp-start 8s steps(1,end) 0s 1 normal both;animation:-amp-start 8s steps(1,end) 0s 1 normal both}@-webkit-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-moz-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-ms-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-o-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}</style><noscript><style amp-boilerplate>body{-webkit-animation:none;-moz-animation:none;-ms-animation:none;animation:none}</style></noscript>
-  <script async src="https://cdn.ampproject.org/v0.js"></script>
-  <script async custom-element="amp-list" src="https://cdn.ampproject.org/v0/amp-list-0.1.js"></script>
-  <script async custom-template="amp-mustache" src="https://cdn.ampproject.org/v0/amp-mustache-latest.js"></script>
-</head>
-<body>
-<article>
-  <h1>amp-list</h1>
-
-  <amp-list width="auto" height="200"
-    layout="fixed-height"
-    src="/test/manual/amp-list-data.json"
-    load-more="auto"
-    load-more-bookmark="next">
-    <template type="amp-mustache">
-      <div class="story-entry">
-        <amp-img src="{{imageUrl}}" width="80" height="60"></amp-img>
-        <a href="/link.html">{{title}}</a>
-      </div>
-    </template>
-    <div overflow>
-      OVERFLOW
-    </div>
-    <div load-more-button>
-      SEE MORE
-    </div>
-    <div load-more-failed>
-      LOAD FAILED
-    </div>
-    <div fallback>
-      FALLBACK
-    </div>
-  </amp-list>
-
-</article>
-</body>
-</html>
