+++ test/manual/amp-list/amp-list-with-ads.amp.html
@@ -72,7 +72,7 @@
 
   <h2>AMP List Basic with Ads</h2>
   <amp-list width="396" height="80" layout="responsive" resizable-children reset-on-refresh
-      src="/test/manual/amp-list-data.json?RANDOM">
+      src="/test/manual/amp-list/amp-list-data.json?RANDOM">
     <template type="amp-mustache">
       <div class="story-entry">
         <amp-img src="{{imageUrl}}" width=80 height=60></amp-img>
