+++ test/manual/amp-list/amp-list.amp.html
@@ -50,7 +50,7 @@
   <h1>amp-list</h1>
 
   <amp-list width="396" height="80" layout=responsive
-      src="/test/manual/amp-list-data.json?RANDOM">
+      src="/test/manual/amp-list/amp-list-data.json?RANDOM">
     <template type="amp-mustache">
       <div class="story-entry">
         <amp-img src="{{imageUrl}}" width=80 height=60></amp-img>
@@ -65,7 +65,7 @@
   <div class="split"></div>
 
   <amp-list width="396" height="80" layout=responsive
-      src="/test/manual/amp-list-data.json?RANDOM">
+      src="/test/manual/amp-list/amp-list-data.json?RANDOM">
     <template type="amp-mustache">
       <div class="story-entry">
         <amp-img src="{{imageUrl}}" width=80 height=60></amp-img>
